Fix-up for the fix for CVE-2013-6396
We pulled an early version of the patch, which had two issues:
 - Tracebacking with EBADF
 - Not matching wildcard hostnames

This patch pulls the upstream patch #18 from the abandoned branch
using pyopenssl, but drops the kludging around close(), which only
makes EBADF worse. Upstream since went to use python-requests, which
presents its own issues (does not support 100-continue for errors).

Only posting this into Rawhide, so we have a quick-fix, but not
stepping on toes of active maintainer, Jakub. Something similar
is necessary for F20 and EPEL. Or not similar. But something.